Refactor import paths and file names for consistency - general helpers
