+++ components/shared/workCard/helper.ts
@@ -2,7 +2,7 @@ import { flatten } from "lodash"
 
 import { Cover } from "@/lib/cover-service-api/model"
 import { SearchWithPaginationQuery, WorkTeaserFragment } from "@/lib/graphql/generated/fbi/graphql"
-import { filterFalsyValuesFromArray } from "@/lib/helpers/helper"
+import { filterFalsyValuesFromArray } from "@/lib/helpers/general"
 
 import { CoverImageUrls } from "./../../../lib/cover-service-api/model/coverImageUrls"
 import { GeneralMaterialTypeCode } from "./../../../lib/graphql/generated/fbi/graphql"
