+++ lib/helpers/general.ts
